+++ code/global/LoggerExtension.cs
@@ -15,5 +15,17 @@ namespace TTTReborn.Globals
 
             log.Info($"[DEBUG][{host}] {obj}");
         }
+
+        public static void Debug(this Logger log, params object[] obj)
+        {
+            if (!Gamemode.Game.Instance.Debug)
+            {
+                return;
+            }
+
+            string host = Host.IsServer ? "SERVER" : "CLIENT";
+
+            log.Info($"[DEBUG][{host}] {string.Join(',', obj)}");
+        }
     }
 }
