+++ code/items/perks/BodyArmor.cs
@@ -4,6 +4,7 @@ namespace TTTReborn.Items
 {
     [Library("ttt_perk_bodyarmor")]
     [Buyable(Price = 100)]
+    [Perk]
     [Hammer.Skip]
     public partial class BodyArmor : Perk
     {
